requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
tiktoken==0.5.1
